Set operations for released vs. pre-released versions
Eliminating pre-release versions from consideration is a common use-case,
so it's good to have that as a first-class feature. The predefined Set
"Released" is like "All" but it excludes versions that have a non-empty
pre-release segment, and so it can be used with the Intersection operator
to reduce a set of candidate versions to exclude pre-releases.