New concept of "Requested" set items
Rather than an assortment of special cases to try to deal with excluding
pre-releases for consideration, here we centralize the concept that a
version set can "request" specific versions, which is distinct from merely
being a member of the set, and retains the individual identities of
"exact" version constraints (Only and Selection) even when they are
combined with infinite sets.

Requests still participate properly in set operations, such that the
requested versions for a set are always a subset of that set. For example,
if an infinite set is subtracted from an exact set then any exact items
within the infinite set are no longer requested in the resulting set.