MeetingConstraints filters out unreleased versions by default
The intent of an expression like >=1.0.0 <2.0.0 is to match all released
versions whose major release number is 1, even though 2.0.0-beta1 is
less than 2.0.0. It's much rarer to want the "strict" interpretation that
would allow 2.0.0-beta1 to match that constraint.

With that in mind, MeetingConstraints now applies that assumption by
default, with the previous behavior now available as
MeetingConstraintsExact.

In practice this means an extra layer of indirection in the sets returned
by MeetingConstraints, but we place the "Released" set first in our
intersection so that we can short-circuit quickly for unreleased versions
and not proceed to evaluate the rest of the possibly-complex expression.