versions/constraints: Tolerate missing space in ruby-like constraint
All of the examples in the rubygems docs show constraints like "~> 1.0.0"
but in practice that space isn't required: "~>1.0.0" is accepted too.
Since some folks seem to prefer writing it that way, we'll allow it here.

Note that this function is still rather more opinionated than rubygems's
requirements parser about things that there's no practical reason to
allow, such as multiple spaces  ("~>  1.0.0"). This module doesn't promise
to be exactly compatible with rubygems, just to follow the spirit of
the rubygems syntax as used idiomatically.