versions: Requested versions eligible only if they were already present
The versions.MeetingConstraints method tried to permit selection of
any prereleases explicitly requested by an exact version constraint, but
it didn't also check that those prereleases were included in the original
set.

That worked okay for positive exact version constraints, like
"1.0.0-beta.1", but was incorrect for negative ones like "! 1.0.0-beta.1"
which still get counted as "requested" in certain cases but that should
not cause them to then be included in the overall result set.

Now we'll filter the list of requested versions first by membership of
the "exact" set, so that the requested status of a version can only
return versions that would've been eligible for inclusion in the original
unfiltered set.