constraints: Catch "v" prefixes on version specs in more cases
Previously we could typically only catch "v" prefixes when written on
operator-less specs, because otherwise it would cause the "sep" production
in the grammar to fail and lead to a misleading error message about
extraneous spaces.

Now we'll accept "v" as part of the "sep" production and then catch it as
a special case in the higher-level parsing code, so we can now report it
even when it appears after an operator.